ai-bank-assistance/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
python-dotenv
sentence-transformers
huggingface-hub
langchain-huggingface
